Refs #20765: Make .h files .hpp
