+++ src/cpp/fastdds/builtin/type_lookup_service/TypeLookupManager.cpp
@@ -38,7 +38,7 @@
 
 #include <rtps/builtin/BuiltinProtocols.h>
 #include <rtps/participant/RTPSParticipantImpl.h>
-#include <rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 #include <rtps/RTPSDomainImpl.hpp>
 
 namespace eprosima {
