+++ src/cpp/rtps/builtin/discovery/endpoint/EDPSimpleListeners.cpp
@@ -35,7 +35,7 @@
 #include <rtps/builtin/discovery/endpoint/EDPSimple.h>
 #include <rtps/builtin/discovery/participant/PDPSimple.h>
 #include <rtps/network/NetworkFactory.h>
-#include <rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 
 using ParameterList = eprosima::fastdds::dds::ParameterList;
