+++ src/cpp/rtps/builtin/discovery/endpoint/EDPUtils.hpp
@@ -31,7 +31,7 @@
 #include <rtps/history/PoolConfig.h>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/participant/RTPSParticipantImpl.h>
-#include <rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 
 #include <memory>
 #include <string>
