+++ src/cpp/rtps/builtin/liveliness/WLP.cpp
@@ -37,7 +37,7 @@
 #include <rtps/builtin/liveliness/WLPListener.h>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/participant/RTPSParticipantImpl.h>
-#include <rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 #include <rtps/resources/TimedEvent.h>
 #include <rtps/writer/LivelinessManager.hpp>
 #include <utils/TimeConversion.hpp>
