+++ src/cpp/rtps/reader/StatefulPersistentReader.cpp
@@ -17,7 +17,7 @@
  *
  */
 
-#include <rtps/reader/StatefulPersistentReader.h>
+#include <rtps/reader/StatefulPersistentReader.hpp>
 
 #include <fastdds/rtps/history/ReaderHistory.h>
