+++ src/cpp/rtps/reader/StatelessPersistentReader.cpp
@@ -17,7 +17,7 @@
  *
  */
 
-#include <rtps/reader/StatelessPersistentReader.h>
+#include <rtps/reader/StatelessPersistentReader.hpp>
 
 #include <fastdds/rtps/history/ReaderHistory.h>
