+++ test/unittest/rtps/history/ReaderHistoryTests.cpp
@@ -22,7 +22,7 @@
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/utils/TimedMutex.hpp>
 #include <rtps/common/ChangeComparison.hpp>
-#include <rtps/reader/StatefulReader.h>
+#include <rtps/reader/StatefulReader.hpp>
 
 
 using namespace eprosima::fastrtps;
